pyproject.toml
[project]
name = "takon-mcp-core"
version = "0.1.0"
description = "MCP Server for Construction and Engineering Management (Daily Reports, CAD, Door Schedules)"
readme = "README.md"
authors = [
    { name = "Takon", email = "takon@example.com" }
]
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.2",
    "ezdxf>=1.4.3",
    "jinja2>=3.1.6",
    "mcp>=1.27.0",
    "moondream>=1.1.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pillow>=10.4.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

[project.scripts]
takon-construction = "takon_mcp_core.construction:main"
takon-cad = "takon_mcp_core.cad:main"

[tool.uv]
managed = true
package = true
